+++ src/Template/TwigRenderer.php
@@ -16,24 +16,34 @@ namespace Rougin\Slytherin\Template;
 class TwigRenderer implements RendererInterface
 {
     /**
-     * @var \Twig_Environment
+     * @var \Twig\Environment
      */
     protected $twig;
 
     /**
-     * @param \Twig_Environment     $twig
-     * @param array<string, string> $globals -- NOTE: To be removed in v1.0.0.
+     * @param \Twig\Environment     $twig
+     * @param array<string, string> $globals (NOTE: To be removed in v1.0.0).
      */
-    public function __construct(\Twig_Environment $twig, array $globals = array())
+    public function __construct($twig, array $globals = array())
     {
-        // NOTE: To be removed in v1.0.0. Use __call instead ---
+        $this->twig = $twig;
+
         foreach ($globals as $key => $value)
         {
-            $twig->addGlobal($key, $value);
+            $this->addGlobal($key, $value);
         }
-        // -----------------------------------------------------
+    }
 
-        $this->twig = $twig;
+    /**
+     * @param  string $name
+     * @param  mixed  $value
+     * @return self
+     */
+    public function addGlobal($name, $value)
+    {
+        $this->twig->addGlobal($name, $value);
+
+        return $this;
     }
 
     /**
@@ -41,28 +51,10 @@ class TwigRenderer implements RendererInterface
      *
      * @param  string               $template
      * @param  array<string, mixed> $data
-     * @param  string               $extension
      * @return string
      */
-    public function render($template, array $data = array(), $extension = 'twig')
+    public function render($template, array $data = array())
     {
-        $file = $template . '.' . $extension;
-
-        return $this->twig->render($file, $data);
-    }
-
-    /**
-     * Calls methods from the Twig instance.
-     *
-     * @param  string  $method
-     * @param  mixed[] $parameters
-     * @return mixed
-     */
-    public function __call($method, $parameters)
-    {
-        /** @var callable */
-        $class = array($this->twig, $method);
-
-        return call_user_func_array($class, $parameters);
+        return $this->twig->render($template, $data);
     }
 }