+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to `Slytherin` will be documented in this file.
 - Checking of available instances in `Container::value`
 - If `ServerRequestInterface` is an argument with a middleware
 - Backward compatibility for `LeagueContainer::set`
+- Backward compatibility for `TwigRenderer::render`
 
 ## [0.9.6](https://github.com/rougin/slytherin/compare/v0.9.5...v0.9.6) - 2023-11-16